refactor: merge container and midwayContainer
