+++ packages/core/test/context/container.test.ts
@@ -1,4 +1,4 @@
-import { Container } from '../../src/context/container';
+import { MidwayContainer as Container } from '../../src/context/midwayContainer';
 import { expect } from 'chai';
 import {
   Grandson,
