+++ packages/decorator/src/common/decoratorManager.ts
@@ -1,17 +1,24 @@
 import 'reflect-metadata';
-import { ObjectDefinitionOptions, TagClsMetadata } from '../interface';
 import {
+  ObjectDefinitionOptions,
+  ObjectIdentifier,
+  ReflectResult,
+  TagClsMetadata,
+  TagPropsMetadata
+} from '../interface';
+import {
+  CLASS_KEY_CONSTRUCTOR, INJECT_TAG,
   MAIN_MODULE_KEY,
   OBJ_DEF_CLS,
-  PRIVATE_META_DATA_KEY,
-  TAGGED_CLS,
+  PRIVATE_META_DATA_KEY, TAGGED,
+  TAGGED_CLS, TAGGED_PROP,
 } from './constant';
-import { classNamed } from './utils';
+
+import { DUPLICATED_INJECTABLE_DECORATOR, DUPLICATED_METADATA, INVALID_DECORATOR_OPERATION } from './errMsg';
+import { Metadata } from './metadata';
+import { getParamNames, classNamed } from '../util';
 
 const debug = require('util').debuglog('decorator:manager');
-
-const STRIP_COMMENTS = /((\/\/.*$)|(\/\*[\s\S]*?\*\/))/gm;
-const ARGUMENT_NAMES = /([^\s,]+)/g;
 
 export type decoratorKey = string | symbol;
 
@@ -721,21 +728,6 @@ export function clearAllModule() {
 }
 
 /**
- * get parameter name from function
- * @param func
- */
-export function getParamNames(func): string[] {
-  const fnStr = func.toString().replace(STRIP_COMMENTS, '');
-  let result = fnStr
-    .slice(fnStr.indexOf('(') + 1, fnStr.indexOf(')'))
-    .match(ARGUMENT_NAMES);
-  if (result === null) {
-    result = [];
-  }
-  return result;
-}
-
-/**
  * get provider id from module
  * @param module
  */
@@ -798,4 +790,204 @@ export function getPropertyType(target, propertyKey: string | symbol) {
 
 export function getMethodReturnTypes(target, propertyKey: string | symbol) {
   return Reflect.getMetadata('design:returntype', target, propertyKey);
+}
+
+
+function _tagParameterOrProperty(
+  metadataKey: string,
+  annotationTarget: any,
+  propertyName: string,
+  metadata: TagPropsMetadata,
+  parameterIndex?: number
+) {
+  let paramsOrPropertiesMetadata: ReflectResult = {};
+  const isParameterDecorator = typeof parameterIndex === 'number';
+  const key: string =
+    parameterIndex !== undefined && isParameterDecorator
+      ? parameterIndex.toString()
+      : propertyName;
+
+  // if the decorator is used as a parameter decorator, the property name must be provided
+  if (isParameterDecorator && propertyName !== undefined) {
+    throw new Error(INVALID_DECORATOR_OPERATION);
+  }
+
+  // read metadata if available
+  if (Reflect.hasOwnMetadata(metadataKey, annotationTarget)) {
+    paramsOrPropertiesMetadata = Reflect.getMetadata(
+      metadataKey,
+      annotationTarget
+    );
+  }
+
+  // get metadata for the decorated parameter by its index
+  let paramOrPropertyMetadata: TagPropsMetadata[] =
+    paramsOrPropertiesMetadata[key];
+
+  if (!Array.isArray(paramOrPropertyMetadata)) {
+    paramOrPropertyMetadata = [];
+  } else {
+    for (const m of paramOrPropertyMetadata) {
+      if (m.key === metadata.key) {
+        throw new Error(`${DUPLICATED_METADATA} ${m.key.toString()}`);
+      }
+    }
+  }
+
+  // set metadata
+  paramOrPropertyMetadata.push(metadata);
+  paramsOrPropertiesMetadata[key] = paramOrPropertyMetadata;
+  Reflect.defineMetadata(
+    metadataKey,
+    paramsOrPropertiesMetadata,
+    annotationTarget
+  );
+}
+
+export function attachConstructorDataOnClass(identifier, clz, type, index) {
+  if (!identifier) {
+    const args = getParamNames(clz);
+    if (clz.length === args.length && index < clz.length) {
+      identifier = args[index];
+    }
+  }
+
+  // save constructor index on class
+  let constructorMetaValue = getClassMetadata(CLASS_KEY_CONSTRUCTOR, clz);
+  if (!constructorMetaValue) {
+    constructorMetaValue = {};
+  }
+  constructorMetaValue[index] = {
+    key: identifier,
+    type,
+  };
+  saveClassMetadata(CLASS_KEY_CONSTRUCTOR, constructorMetaValue, clz);
+}
+
+interface InjectOptions {
+  identifier: ObjectIdentifier;
+  target: any;
+  targetKey: string;
+  index?: number;
+  args?: any;
+}
+/**
+ * 构造器注入
+ * @param opts 参数
+ */
+export function saveConstructorInject(opts: InjectOptions) {
+  let identifier = opts.identifier;
+  if (!identifier) {
+    const args = getParamNames(opts.target);
+    if (opts.target.length === args.length && opts.index < opts.target.length) {
+      identifier = args[opts.index];
+    }
+  } else if (identifier.includes('@') && !identifier.includes(':')) {
+    const args = getParamNames(opts.target);
+    if (opts.target.length === args.length && opts.index < opts.target.length) {
+      identifier = `${identifier}:${args[opts.index]}`;
+    }
+  }
+  const metadata = new Metadata(INJECT_TAG, identifier);
+  metadata.args = opts.args;
+  _tagParameterOrProperty(
+    TAGGED,
+    opts.target,
+    opts.targetKey,
+    metadata,
+    opts.index
+  );
+}
+
+export function getConstructorInject(target: any): TagPropsMetadata[] {
+  return Reflect.getMetadata(TAGGED, target);
+}
+/**
+ * 属性注入
+ * @param opts 参数
+ */
+export function savePropertyInject(opts: InjectOptions) {
+  let identifier = opts.identifier;
+  if (!identifier) {
+    identifier = opts.targetKey;
+  }
+  if (identifier.includes('@') && !identifier.includes(':')) {
+    identifier = `${identifier}:${opts.targetKey}`;
+  }
+  const metadata = new Metadata(INJECT_TAG, identifier);
+  metadata.args = opts.args;
+  _tagParameterOrProperty(
+    TAGGED_PROP,
+    opts.target.constructor,
+    opts.targetKey,
+    metadata
+  );
+}
+
+export function getPropertyInject(target: any): TagPropsMetadata[] {
+  return Reflect.getMetadata(TAGGED_PROP, target);
+}
+/**
+ * class 元数据定义
+ * @param target class
+ * @param props 属性
+ */
+export function saveObjectDefProps(target: any, props: object = {}) {
+  if (Reflect.hasMetadata(OBJ_DEF_CLS, target)) {
+    const originProps = Reflect.getMetadata(OBJ_DEF_CLS, target);
+
+    Reflect.defineMetadata(
+      OBJ_DEF_CLS,
+      Object.assign(originProps, props),
+      target
+    );
+  } else {
+    Reflect.defineMetadata(OBJ_DEF_CLS, props, target);
+  }
+  return target;
+}
+
+export function getObjectDefProps(target: any): ObjectDefinitionOptions {
+  return Reflect.getMetadata(OBJ_DEF_CLS, target);
+}
+/**
+ * class provider id
+ * @param identifier id
+ * @param target class
+ * @param override 是否覆盖
+ */
+export function saveProviderId(
+  identifier: ObjectIdentifier,
+  target: any,
+  override?: boolean
+) {
+  if (Reflect.hasOwnMetadata(TAGGED_CLS, target) && !override) {
+    throw new Error(DUPLICATED_INJECTABLE_DECORATOR);
+  }
+
+  if (!identifier) {
+    identifier = classNamed(target.name);
+  }
+
+  Reflect.defineMetadata(
+    TAGGED_CLS,
+    {
+      id: identifier,
+      originName: target.name,
+    },
+    target
+  );
+
+  if (!Reflect.hasMetadata(OBJ_DEF_CLS, target)) {
+    Reflect.defineMetadata(OBJ_DEF_CLS, {}, target);
+  }
+
+  return target;
+}
+/**
+ * 是否使用了 saveProviderId
+ * @param target class
+ */
+export function isProvide(target: any): boolean {
+  return Reflect.hasOwnMetadata(TAGGED_CLS, target);
 }
