+++ packages/decorator/src/index.ts
@@ -5,7 +5,6 @@ export * from './common/decoratorManager';
 export * from './common/errMsg';
 export * from './common/metadata';
 export * from './common/scopeEnum';
-export * from './common/utils';
 export * from './faas/fun';
 export * from './faas/handler';
 export * from './web/requestMapping';
